# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "basecast-get-data",
    "description": "Serves the basecast lake, the processed tables and the pipeline runs to basecast-app. Every route but /health needs `Authorization: Bearer <API_TOKEN>`.",
    "version": "0.1.0"
  },
  "paths": {
    "/lake/sources": {
      "get": {
        "tags": [
          "lake"
        ],
        "summary": "Lake Sources",
        "operationId": "lake_sources_lake_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/lake/list": {
      "get": {
        "tags": [
          "lake"
        ],
        "summary": "Lake List",
        "operationId": "lake_list_lake_list_get",
        "parameters": [
          {
            "name": "prefix",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Folder to list, ending in '/', or '' for the bucket root",
              "default": "",
              "title": "Prefix"
            },
            "description": "Folder to list, ending in '/', or '' for the bucket root"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Keep only names containing this text (case-insensitive)",
              "default": "",
              "title": "Q"
            },
            "description": "Keep only names containing this text (case-insensitive)"
          },
          {
            "name": "recursive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Under raw/, list every object below the prefix",
              "default": false,
              "title": "Recursive"
            },
            "description": "Under raw/, list every object below the prefix"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/lake/object": {
      "get": {
        "tags": [
          "lake"
        ],
        "summary": "Lake Object",
        "operationId": "lake_object_lake_object_get",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Object key, e.g. raw/source=ercot_gis/dt=2026-08-01/<file>",
              "title": "Key"
            },
            "description": "Object key, e.g. raw/source=ercot_gis/dt=2026-08-01/<file>"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/lake/object/structure": {
      "get": {
        "tags": [
          "lake"
        ],
        "summary": "Lake Structure",
        "operationId": "lake_structure_lake_object_structure_get",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          },
          {
            "name": "member",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Member"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StructureResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/lake/object/rows": {
      "get": {
        "tags": [
          "lake"
        ],
        "summary": "Lake Rows",
        "operationId": "lake_rows_lake_object_rows_get",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          },
          {
            "name": "member",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A file inside a zip",
              "title": "Member"
            },
            "description": "A file inside a zip"
          },
          {
            "name": "sheet",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Spreadsheet sheet; the first one by default",
              "title": "Sheet"
            },
            "description": "Spreadsheet sheet; the first one by default"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          },
          {
            "name": "with_summary",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Also return the total row count",
              "default": false,
              "title": "With Summary"
            },
            "description": "Also return the total row count"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RowsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/lake/object/text": {
      "get": {
        "tags": [
          "lake"
        ],
        "summary": "Lake Text",
        "operationId": "lake_text_lake_object_text_get",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          },
          {
            "name": "member",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Member"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/lake/object/url": {
      "get": {
        "tags": [
          "lake"
        ],
        "summary": "Lake Signed Url",
        "operationId": "lake_signed_url_lake_object_url_get",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/lake/object/content": {
      "get": {
        "tags": [
          "lake"
        ],
        "summary": "Lake Content",
        "description": "The file's bytes, with HTTP Range support (pdf.js reads PDFs in ranges).",
        "operationId": "lake_content_lake_object_content_get",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          },
          {
            "name": "member",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Member"
            }
          },
          {
            "name": "Range",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Range"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/octet-stream": {}
            }
          },
          "206": {
            "description": "Partial content"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListingResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "$ref": "#/components/schemas/LakeListing"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "ListingResponse"
      },
      "Meta": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "data_as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data As Of"
          },
          "model_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Run Id"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated",
            "default": false
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "generated_at"
        ],
        "title": "Meta"
      },
...
      "ObjectResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "$ref": "#/components/schemas/ObjectDetail"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "ObjectResponse"
      },
...
      "RowsResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "$ref": "#/components/schemas/RowsPage"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "RowsResponse"
      },
...
      "SignedUrlResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "$ref": "#/components/schemas/SignedUrl"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "SignedUrlResponse"
      },
...
      "SourcesResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "$ref": "#/components/schemas/SourcesData"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "SourcesResponse"
      },
      "StructureResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "$ref": "#/components/schemas/ObjectStructure"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "StructureResponse"
      },
...
      "TextResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "data": {
            "$ref": "#/components/schemas/ObjectText"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "TextResponse"
      },
...
    }
  }
}